chore: split tests
